Fix bug around header boundary. Cleanup.

+++ tsk/util/Bitlocker/BitlockerParser.h
@@ -168,8 +168,8 @@ private:
     int decryptSector(TSK_DADDR_T offset, uint8_t* data);
     int decryptSectorAESCBC_noDiffuser(uint64_t offset, uint8_t* data);
     int decryptSectorAESCBC_diffuser(uint64_t offset, uint8_t* data);
-    void decryptDiffuserA(uint8_t* data, size_t dataLen, uint8_t* result);
-    void decryptDiffuserB(uint8_t* data, size_t dataLen, uint8_t* result);
+    void decryptDiffuserA(uint8_t* data, uint16_t dataLen, uint8_t* result);
+    void decryptDiffuserB(uint8_t* data, uint16_t dataLen, uint8_t* result);
     int decryptSectorAESXTS(uint64_t offset, uint8_t* data);
 
     list<uint64_t> m_fveMetadataOffsets;
